+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - Use modern build tools: vite, vitest
 - Show shortcuts in tooltip
 - Wizard: checkbox to include/exclude query comments
+- Migrate to @openstreetmap/id-tagging-schema
 
 ## 2022-01-09